Move Cosmos proxies to own dir
