+++ cosmos/proxies/global-style-proxy.js
